refactor: move simplification into new file
