+++ crates/rnote-compose/src/meson.build
@@ -27,6 +27,7 @@ rnote_compose_sources = files(
     'penpath/element.rs',
     'penpath/mod.rs',
     'penpath/segment.rs',
+    'penpath/simplification.rs',
     'serialize.rs',
     'shapes/arrow.rs',
     'shapes/cubbez.rs',
